refactor: 패키지 depth 줄이기
