--- api/src/main/java/shop/ink3/api/cart/cart/entity/Cart.java
@@ -1,36 +1,0 @@
-package shop.ink3.api.cart.cart.entity;
-
-import jakarta.persistence.Entity;
-import jakarta.persistence.GeneratedValue;
-import jakarta.persistence.GenerationType;
-import jakarta.persistence.Id;
-import jakarta.persistence.Table;
-import lombok.AccessLevel;
-import lombok.AllArgsConstructor;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-
-@Entity
-@Table(name = "carts")
-@Getter
-@NoArgsConstructor(access = AccessLevel.PROTECTED)
-@AllArgsConstructor
-public class Cart {
-
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long id;
-
-    //TODO: 임시
-    private Long userId;
-    //TODO: 임시
-    private Long bookId;
-
-    private int quantity;
-
-    public Cart(Long userId, Long bookId, int quantity) {
-        this.userId = userId;
-        this.bookId = bookId;
-        this.quantity = quantity;
-    }
-}
