+++ api/src/main/java/shop/ink3/api/cart/repository/CartRepository.java
@@ -1,10 +1,10 @@
-package shop.ink3.api.cart.cart.repository;
+package shop.ink3.api.cart.repository;
 
 import java.util.List;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-import shop.ink3.api.cart.cart.entity.Cart;
+import shop.ink3.api.cart.entity.Cart;
 
 public interface CartRepository extends JpaRepository<Cart, Long> {
     List<Cart> findByUserId(Long userId);
